finished new door(UT), Added new universal lock gui
changes are untest as more work is still needed first. Finished the door
though untested yet it is a basic rewrite of the door that adds a
TileEntity, packet sending, and just check the user on right click.
Change the doors user list gui so that both it, the chest, and other
lockable devices can use it. Though the gui itself still needs reworked
and its textures remade.

Also made changes to Dark library to make it stand out as a library in
case one of my other mods is added to repo using the same 'dark' package