Corrections
When your working on the code use Ctrl+shift+O to fix all import on a
class file. As well you can select all files and do the same. If you
going to change the name on something use refractor as it will correct
all name in the entire project. Ctrl+alt+r.